requirements.txt
django==5.2.3
djangorestframework==3.16.0
django-rest-knox==5.0.2
freezegun==1.5.2
pyyaml==6.0.2
requests==2.32.4
